requirements.txt
python-telegram-bot[webhooks]~=20.1
APScheduler~=3.10.1
pytz>=2018.6
# tornado~=6.2
# uvicorn~=0.17.0
# starlette~=0.20.0

colorlog~=6.7.0
aiohttp~=3.9.1
# pyzbar~=0.1.9
# matplotlib~=3.7.1
# html2text>=2020.1.16
# pycryptodome~=3.17
# feedparser~=6.0.10
# influxdb~=5.3.1
# flask~=2.2.3
# gevent>=22.10.2
# Mastodon.py~=1.5
# webdavclient3~=3.14.6
